replaced bsp_pid_control.c/h with bsp_sbv.c/h

+++ Libraries/KEY/bsp_key.c
@@ -193,7 +193,7 @@ void BSP_KEY_UsageDemo(void)
 **/
 uint32_t Key_GetTime(void) 
 {
-	unsigned long count;
+	uint32_t count;
 	get_tick_count(&count);
 	return count;
 }
